Refactored view groupings
